==> code/data/03_ipums_clean.ipynb <==
import pandas as pd
import numpy as np
from pathlib import Path

# 1. Configuration

INPUT_FILE = 'data/IPUMS/ipums_1823.csv'
OUTPUT_DIR = 'data/processed/'

MISSING_CODES = {
    'OWNERSHP': 0,
    'OWNERSHPD': 0,
    'SEX': 9,
    'MARST': 9,
    'HISPAN': 9,
    'EDUC': 99,
    'EMPSTAT': [0, 9],
    'EMPSTATD': [0, 99],
    'MIGRATE': [0, 9],
    'MIGRATE1D': [00, 90],
    'RENT': 9999999,
    'HHINCOME': 9999999,
    'VALUEH': [9999998, 9999999],
    'INCTOT': 9999999
}

# 2. Data loading

dtype_dict = {
        'YEAR': 'int32',
        'SAMPLE': 'int32',
        'SERIAL': 'int64',
        'CBSERIAL': 'int64',
        'STATEFIP': 'int16',
        'COUNTYFIP': 'int16',
        'PUMA': 'int32',
        'GQ': 'int8',
        'PERNUM': 'int16',
    }

df = pd.read_csv('../../data/IPUMS/ipums_1722_FLonly.csv', dtype=dtype_dict)

df.shape

df.sample()

# 2. Cleaning

# Handles NA

for var, codes in MISSING_CODES.items():
    if var in df.columns:
        df[var] = df[var].replace(codes, np.nan)

if 'INCTOT' in df.columns:
    df.INCTOT = df.INCTOT.replace(MISSING_CODES.get('INCTOT', []), np.nan)

# Re-classing numerical columns

num_vars = ['VALUEH', 'HHINCOME', 'RENT', 'HHWT']

for x in num_vars:
    if x in df.columns:
        df[x] = pd.to_numeric(df[x], errors='coerce')

# Homeownership flags

if 'OWNERSHP' in df.columns:
    df['is_owner'] = (df.OWNERSHP == 1).astype(int)
    df['is_renter'] = (df.OWNERSHP == 2).astype(int)

if 'OWNERSHPD' in df.columns:
    df['owns_free_clear'] = (df.OWNERSHPD == 12).astype(int)
    df['owns_with_mortgage'] = (df.OWNERSHPD == 13).astype(int)

df.head()

df.isnull().sum()

df.to_csv('../../data/processed/FL_IPUMS.csv')
